# docs/swagger.yaml
basePath: /api
definitions:
  dto.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UpsertCategoryDto:
    properties:
      type:
        type: string
    required:
    - type
    type: object
  dto.UpsertUserDto:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  models.CategoryModel:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  utils.HttpError:
    properties:
      message: {}
      stack_trace: {}
    type: object
  utils.HttpSuccess-any:
    properties:
      data: {}
      message: {}
    type: object
  utils.HttpSuccess-array_models_CategoryModel:
    properties:
      data:
        items:
          $ref: '#/definitions/models.CategoryModel'
        type: array
      message: {}
    type: object
  utils.HttpSuccess-dto_UpsertUserDto:
    properties:
      data:
        $ref: '#/definitions/dto.UpsertUserDto'
      message: {}
    type: object
  utils.HttpSuccess-models_CategoryModel:
    properties:
      data:
        $ref: '#/definitions/models.CategoryModel'
      message: {}
    type: object
  utils.HttpSuccess-models_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/models.LoginResponse'
      message: {}
    type: object
  utils.HttpSuccess-string:
    properties:
      data:
        type: string
      message: {}
    type: object
host: localhost:3030
info:
  contact:
    email: support@swagger.io
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
  /categories:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-array_models_CategoryModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: get all categories based on user
      tags:
      - Categories
    post:
      parameters:
      - description: Create Category DTO
        in: body
        name: Category
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertCategoryDto'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.HttpSuccess-models_CategoryModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: create a category
      tags:
      - Categories
  /categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    patch:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Category On User
        in: body
        name: SocialMedia
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertCategoryDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-models_CategoryModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
  /user:
    delete:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: delete current user based on JWT
      tags:
      - User
    put:
      parameters:
      - description: Update User Based On Token
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertUserDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-dto_UpsertUserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      security:
      - BearerAuth: []
      summary: create a user
      tags:
      - User
  /user/login:
    post:
      parameters:
      - description: Login User DTO
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.HttpSuccess-models_LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: login a user
      tags:
      - User
  /user/register:
    post:
      parameters:
      - description: Create User DTO
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpsertUserDto'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.HttpSuccess-dto_UpsertUserDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HttpError'
      summary: create a user
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"